Remove `include/grpcpp/impl/codegen/byte_buffer.h`
