+++ include/grpcpp/impl/codegen/method_handler.h
@@ -21,10 +21,10 @@
 
 // IWYU pragma: private, include <grpcpp/support/method_handler.h>
 
-#include <grpcpp/impl/codegen/byte_buffer.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
 #include <grpcpp/impl/codegen/rpc_service_method.h>
 #include <grpcpp/impl/codegen/sync_stream.h>
+#include <grpcpp/support/byte_buffer.h>
 
 namespace grpc {
