+++ include/grpcpp/impl/codegen/rpc_service_method.h
@@ -28,10 +28,10 @@
 #include <vector>
 
 #include <grpc/impl/codegen/log.h>
-#include <grpcpp/impl/codegen/byte_buffer.h>
 #include <grpcpp/impl/codegen/config.h>
 #include <grpcpp/impl/codegen/rpc_method.h>
 #include <grpcpp/impl/codegen/status.h>
+#include <grpcpp/support/byte_buffer.h>
 
 namespace grpc {
 class ServerContextBase;
